Samenvoeging_klassen
Addressbook is vervangen door klantbestand. Deze implementeert klantDAO
die nu een interface is. adreslijst is hetzelfde gebleven. Bestellijst
en bestellingDAO zijn ook samengevoegd en dit is bestellijst gebleven.
De oude bestellingDAO is vervangen door de bestellingenDAO van Maurice.
Er zijn nog methdoes die we aan moeten passen omdat ik met klassen had
gewerkt. We moeten hier met klant_id, bestelling_id, adres_id etc. gaan
werken en eerst de read methode aanroepen om verder te werken met de
objecten.
