+++ Bestellijst.java
@@ -1,3 +1,7 @@
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.math.BigDecimal;
 import java.sql.*;
 
 import com.sun.rowset.*;
@@ -8,9 +12,15 @@ import javax.sql.rowset.JdbcRowSet;
 
 
 public class Bestellijst implements BestellingenDAO {
+	
+	private final static String URL = "jdbc:mysql://localhost/Adresboek";
+	private final static String USERNAME = "root";
+	private final static String PASSWORD = "komt_ie";
+	
 	private static Connection connection;
 	
-	public Bestellijst(){
+	public Bestellijst() {
+		// Ensure the connection is initialized
 		try {
 			getConnection();
 			
@@ -19,23 +29,22 @@ public class Bestellijst implements BestellingenDAO {
 		}
 	}
 	
-	public static Connection getConnection(String user, String password) throws SQLException {
-		if (connection == null || connection.isClosed()){
-			try {
+	
+	public static Connection getConnection() throws SQLException{
+		if (connection == null || connection.isClosed())
+			try{
 				Class.forName("com.mysql.jdbc.Driver");
 				System.out.print("Driver loaded... ");
-			
-				connection = DriverManager.getConnection("jdbc:mysql://localhost/Adresboek", user, password);
+				connection = DriverManager.getConnection("jdbc:mysql://localhost/Adresboek", "root", "komt_ie");
 				System.out.println("Database connected!");
-				
-			} catch (ClassNotFoundException e) {
+			
+			}catch (ClassNotFoundException e){
 				e.printStackTrace();
 			}
-		}
-				
 		return connection;
 	}
-	
+				
+		
 	@Override//Het was volgens mij de bedoeling dat iedere keer als er ingelogd wordt er een nieuwe tabel wordt aangemaakt
 	public void maakTabel() throws SQLException {
 		getConnection();
@@ -62,31 +71,27 @@ public class Bestellijst implements BestellingenDAO {
 		
 		System.out.println("table Bestellingen created!");
 	}
-
-	@Override
-	public void voegBestellingToe(Bestelling bestelling) throws SQLException {
-		getConnection();
-		
-		ArrayList<Artikel> artikelen = bestelling.getArtikelen();
-		
+	// DIT NOG EVEN NAKIJKEN KLOPT NU NOG NIET... MOET WORDEN KLANT_ID EN BESTELLING
+	public void voegBestellingToe(Klant klant, Bestelling bestelling) throws SQLException {
 		String sql = "";
 		String values = "";
+		Connection connection = DatabaseConnection.getConnection();
 		
-		for (int i = 1; i <= Math.min(3, artikelen.size()); i++){
-			sql += String.format(", artikel%d_id, artikel%d_naam, artikel%d_aantal, artikel%d_prijs", i, i, i, i);
+		for (int i = 0; i < Math.min(3, bestelling.artikelen.size()); i++){
+			sql += String.format(", artikel%d_id, artikel%d_naam, artikel%d_aantal, artikel%d_prijs", i + 1, i + 1, i + 1, i + 1);
 			values += ", ?, ?, ?, ?";
 		}
 				
 		PreparedStatement voegToe = connection.prepareStatement(
 				String.format("INSERT INTO Bestelling (klant_id%s) VALUES (?%s)", sql, values), Statement.RETURN_GENERATED_KEYS);
 		
-		voegToe.setInt(1, bestelling.getKlant_id());
+		voegToe.setInt(1, klant.getId());
 		
-		for (int i = 0; i < Math.min(3, artikelen.size()); i++){
-			voegToe.setInt(2 + i * 4, artikelen.get(i).getId());
-			voegToe.setString(3 + i * 4, artikelen.get(i).getNaam());
-			voegToe.setInt(4 + i * 4, artikelen.get(i).getAantal());
-			voegToe.setDouble(5 + i * 4, artikelen.get(i).getPrijs());
+		for (int i = 0; i < Math.min(3, bestelling.artikelen.size()); i++){
+			voegToe.setInt(2 + i * 4, bestelling.artikelen.get(i).getId());
+			voegToe.setString(3 + i * 4, bestelling.artikelen.get(i).getNaam());
+			voegToe.setInt(4 + i * 4, bestelling.artikelen.get(i).getAantal());
+			voegToe.setBigDecimal(5 + i * 4, bestelling.artikelen.get(i).getPrijs());
 		}	
 		voegToe.executeUpdate();
 		
@@ -94,7 +99,7 @@ public class Bestellijst implements BestellingenDAO {
 		
 		if (rs.isBeforeFirst()){
 			rs.next();
-			bestelling.setId(rs.getInt(1));
+			bestelling.setBestelling_id(rs.getInt(1));
 		}
 			
 		System.out.println("Bestelling added!");
@@ -102,32 +107,72 @@ public class Bestellijst implements BestellingenDAO {
 
 	public static Bestelling haalBestelling(int id) throws SQLException {
 		getConnection();
+		Bestelling bestelling = new Bestelling();
+		String haalBestelling = "SELECT artikel_id, artikel_naam, artikel_prijs, artikel_aantal FROM Bestelling WHERE bestelling_id = ?";
 		
-		PreparedStatement haalBestelling = connection.prepareStatement
-				("SELECT * FROM Bestelling WHERE bestelling_id = ?");
-		haalBestelling.setInt(1, id);
-		ResultSet gegevens = haalBestelling.executeQuery();
+		RowSet rowSet = new JdbcRowSetImpl();
+		rowSet.setUrl(URL);
+		rowSet.setPassword(PASSWORD);
+		rowSet.setUsername(USERNAME);
+		rowSet.setCommand(haalBestelling);
+		rowSet.setInt(1, id);
 		
-		JdbcRowSet rowSet = new JdbcRowSetImpl(gegevens);
+		rowSet.execute();	
 		
-		Bestelling bestelling = new Bestelling(rowSet.getInt(2));
-		
-		for (int i = 0; i < 3; i++){
-			if (rowSet.getInt(3 + i * 4) > 0){
-				bestelling.voegArtikelToe(new Artikel(rowSet.getInt(3 + i * 4), rowSet.getString(4 + i * 4), rowSet.getInt(5 + i * 4), rowSet.getDouble(6 + i * 4)));
+		while (rowSet.next()){
+			
+			for (int i = 0; i < 3; i++){
+					bestelling.voegArtikelToeAanBestelling(new Artikel(rowSet.getInt(i * 4 + 1), rowSet.getString(i * 4 + 2), rowSet.getBigDecimal(i * 4 + 3), rowSet.getInt(i * 4 + 4)));
 			}
 		}
 		rowSet.close();
 		
-		bestelling.toString();//misschien beter niet hier?!?
-		
 		return bestelling;
 	}
 	
+	public void haalKlantEnBestelling(int bestelling_id) throws SQLException{
+		RowSet rowSet = new JdbcRowSetImpl();
+		String query = "Select * from klant INNER JOIN bestelling ON klant.klant_id=bestelling.klant_id where bestelling_id=?;";
+		ResultSetMetaData rsMD = null;
+		try{
+			rowSet.setUrl(URL);
+			rowSet.setPassword(PASSWORD);
+			rowSet.setUsername(USERNAME);
+			rowSet.setCommand(query);
+			rowSet.setInt(1, bestelling_id);
+			
+			rowSet.execute();
+			
+			if (!rowSet.isBeforeFirst()){
+				System.out.println("no results found for this bestelling_id");
+				return;
+			}
+			else{
+				rsMD = rowSet.getMetaData();
+				for (int i = 1; i <= rsMD.getColumnCount(); i++){
+					System.out.printf("%-12s\t", rsMD.getColumnLabel(i));
+				}
+				System.out.println();
+				while(rowSet.next()){
+					for (int i = 1; i <= rsMD.getColumnCount(); i++){
+						System.out.printf("%-12s\t", rowSet.getObject(i));
+						
+						if (i % rsMD.getColumnCount() == 0){
+							System.out.println();
+						}
+					}
+				}
+			}
+		}
+			finally{
+				rowSet.close();
+			}
+		}
+			
 	@Override
 	public ArrayList<Bestelling> haalBestellijst() throws SQLException{
 		getConnection();
-		
+				
 		ArrayList<Bestelling> bestellijst = new ArrayList<>();
 				
 		RowSet rowSet = new JdbcRowSetImpl(connection);
@@ -154,29 +199,57 @@ public class Bestellijst implements BestellingenDAO {
 	}
 	
 	@Override
-	public void updateBestelling(Bestelling bestelling) throws SQLException{
-		getConnection();
+public void updateBestelling(Bestelling bestelling) throws SQLException{
 		
-		ArrayList<Artikel> artikelen = bestelling.getArtikelen();
+		Connection connection = null;
+		PreparedStatement updateBestelling= null;
+		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+		connection = DatabaseConnection.getConnection();
 		
-		String sqlupdate = "UPDATE Bestelling SET ";
-		for (int i = 0; i < Math.min(3, artikelen.size()); i++){
-			sqlupdate += String.format
-					("artikel%d_id = ?, artikel%d_naam = ?, artikel%d_aantal = ?, artikel%d_prijs = ? ", i +1, i +1, i +1, i +1);
+		for (int i = 1; i <= Math.min(3, bestelling.artikelen.size()); i++){
+			
+			int bestellingId = bestelling.getBestelling_id();
+			String updateBestellingQuery = String.format("UPDATE bestelling SET artikel%d_naam=?, artikel%d_prijs=?, artikel%d_aantal=?, artikel%d_id=? WHERE bestelling_id=?" , i, i, i, i);
+				
+		try {
+			BigDecimal artikelPrijs = new BigDecimal(0);
+			System.out.println("enter new artikel naam");
+			String artikelNaam = br.readLine();
+			bestelling.artikelen.get(i).setNaam(artikelNaam);
+			System.out.println("enter new prijs");
+			String artikelPrijsStr = br.readLine();
+			artikelPrijs = artikelPrijs.add(new BigDecimal(artikelPrijsStr));
+			bestelling.artikelen.get(i).setPrijs(artikelPrijs);
+			System.out.println("enter new aantal");
+			String artikelAantalStr = br.readLine();
+			int artikelAantal = Integer.parseInt(artikelAantalStr);
+			bestelling.artikelen.get(i).setAantal(artikelAantal);
+			System.out.println("enter new artikel id");
+			String artikelIdStr = br.readLine();
+			int artikelId = Integer.parseInt(artikelIdStr);
+			bestelling.artikelen.get(i).setId(artikelId);
+			
+			updateBestelling = connection.prepareStatement(updateBestellingQuery);
+			updateBestelling.setString(1, artikelNaam);
+			updateBestelling.setBigDecimal(2, artikelPrijs);
+			updateBestelling.setInt(3, artikelAantal);
+			updateBestelling.setInt(4, artikelId);
+			updateBestelling.setInt(5, bestellingId);
+			
+			
+			updateBestelling.executeUpdate();
 		}
-		sqlupdate += "WHERE bestelling_id = " + bestelling.getId();
-		
-		PreparedStatement update = connection.prepareStatement(sqlupdate);
-		
-		for (int i = 0; i < Math.min(3, artikelen.size()); i++){
-			update.setInt(1 + i * 4, artikelen.get(i).getId());
-			update.setString(2 + i * 4, artikelen.get(i).getNaam());
-			update.setInt(3 + i * 4, artikelen.get(i).getAantal());
-			update.setDouble(4 + i * 4, artikelen.get(i).getPrijs());
+		catch (IOException ex){
+			ex.printStackTrace();
 		}
-		update.execute();
+		catch (SQLException ex){
+			ex.printStackTrace();
+		}
 		
-		System.out.println("Bestelling " + bestelling.getId() + " is veranderd!");
+	}
+			updateBestelling.close();
+			connection.close();
+		
 	}
 	
 	@Override
